requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.1
docx==0.2.4
etelemetry==0.3.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
fitz==0.0.1.dev2
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.6
isodate==0.6.1
looseversion==1.3.0
lxml==5.1.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
prov==2.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
traits==6.3.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
